+++ app/models/marty/grid_index_int4range.rb
@@ -1,3 +1,0 @@
-class Marty::GridIndexInt4range < Marty::Base
-  validates :created_dt, :data_grid_id, :attr, :index, presence: true
-end