+++ app/models/marty/grid_index_string.rb
@@ -1,3 +1,0 @@
-class Marty::GridIndexString < Marty::Base
-  validates :created_dt, :data_grid_id, :attr, :index, presence: true
-end